requirements.txt
opencv-python>=4.3.0.36
numpy>=1.19.0
Flask>=1.1.2
pillow>=8.4
scipy>=1.7.3